test: copy test package to run linter outside of this project
test package has no dependencies except for eslint-plugin-editorconfig, but without copying the test package outside of this project, the test package is affected by node_modules under the project root.

Before this change, eslint-plugin-typescript did not work without typescript package, but I could not be noticed about it because typescript package is already installed in the node_modules under the project root although typescript is not installed in the test package itself.